requirements-prod.txt
Django==5.2.7
Pillow==12.0.0
phonenumbers==9.0.17
django-phonenumber-field==8.3.0
babel==2.17.0
setuptools==80.9.0
requests~=2.32.5
python-dotenv==1.2.1
colorama~=0.4.6
django-q2==1.8.0
icalendar~=6.3.1
gunicorn==21.2.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
dj-database-url==2.1.0

